Implemnted the test cases with the updated build file and ivy.xml

+++ Homework10/src/main/umbcs680/hw10/PriceComparator.java
@@ -7,6 +7,6 @@ public class PriceComparator implements Comparator<Car> {
 
     @Override
     public int compare(Car car1, Car car2) {
-        return Float.compare(car1.getPrice(), car2.getPrice());
+        return Double.compare(car1.getPrice(), car2.getPrice());
     }
 }
